[#16]REFACTOR: refactoring complete - synchronizing with algorithm_dev server
