+++ scripts/core_algos/assets.py
